fix(integration-upwork): prevent memory leaks in lazy-loaded modules
