+++ packages/plugins/integration-upwork-ui/src/lib/integration-upwork-ui.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { filter, merge, tap } from 'rxjs';
 import {
 	NbActionsModule,
 	NbButtonModule,
@@ -16,12 +15,8 @@ import {
 	NbToggleModule,
 	NbTooltipModule
 } from '@nebular/theme';
-import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
-import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { LanguagesEnum } from '@gauzy/contracts';
-import { distinctUntilChange } from '@gauzy/ui-core/common';
-import { Store } from '@gauzy/ui-core/core';
-import { HttpLoaderFactory, I18nService } from '@gauzy/ui-core/i18n';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { HttpLoaderFactory } from '@gauzy/ui-core/i18n';
 import {
 	SmartDataViewLayoutModule,
 	SelectorsModule,
@@ -36,9 +31,11 @@ import { TransactionsComponent } from './components/transactions/transactions.co
 import { ContractsComponent } from './components/contracts/contracts.component';
 import { SyncDataSelectionComponent } from './components/sync-data-selection/sync-data-selection.component';
 import { ReportsComponent } from './components/reports/reports.component';
+import { IntegrationUpworkLayoutComponent } from './integration-upwork.layout.component';
 
 @NgModule({
 	declarations: [
+		IntegrationUpworkLayoutComponent,
 		UpworkComponent,
 		UpworkAuthorizeComponent,
 		TransactionsComponent,
@@ -75,32 +72,4 @@ import { ReportsComponent } from './components/reports/reports.component';
 		TableComponentsModule
 	]
 })
-@UntilDestroy()
-export class IntegrationUpworkUiModule {
-	constructor(
-		readonly _translateService: TranslateService,
-		readonly _store: Store,
-		readonly _i18nService: I18nService
-	) {
-		this.initializeUiLanguagesAndLocale(); // Initialize UI languages and Update Locale
-		console.log(`integration upwork ui module plugin initialized`);
-	}
-
-	/**
-	 * Initialize UI languages and Update Locale
-	 */
-	private initializeUiLanguagesAndLocale() {
-		// Observable that emits when preferred language changes.
-		const preferredLanguage$ = merge(this._store.preferredLanguage$, this._i18nService.preferredLanguage$).pipe(
-			distinctUntilChange(),
-			filter((lang: string | LanguagesEnum) => !!lang),
-			tap((lang: string | LanguagesEnum) => {
-				this._translateService.use(lang);
-			}),
-			untilDestroyed(this)
-		);
-
-		// Subscribe to initiate the stream
-		preferredLanguage$.subscribe();
-	}
-}
+export class IntegrationUpworkUiModule {}
