+++ packages/plugins/integration-upwork-ui/src/lib/integration-upwork.routes.ts
@@ -5,51 +5,58 @@ import { UpworkAuthorizeComponent } from './components/upwork-authorize/upwork-a
 import { TransactionsComponent } from './components/transactions/transactions.component';
 import { ContractsComponent } from './components/contracts/contracts.component';
 import { ReportsComponent } from './components/reports/reports.component';
+import { IntegrationUpworkLayoutComponent } from './integration-upwork.layout.component';
 
 @NgModule({
 	imports: [
 		RouterModule.forChild([
 			{
 				path: '',
-				component: UpworkAuthorizeComponent,
-				data: { state: true }
-			},
-			{
-				path: 'regenerate',
-				component: UpworkAuthorizeComponent,
-				data: { state: false }
-			},
-			{
-				path: ':id',
-				component: UpworkComponent,
+				component: IntegrationUpworkLayoutComponent,
 				children: [
 					{
 						path: '',
-						redirectTo: 'contracts',
-						pathMatch: 'full'
+						component: UpworkAuthorizeComponent,
+						data: { state: true }
 					},
 					{
-						path: 'activities',
-						component: TransactionsComponent
+						path: 'regenerate',
+						component: UpworkAuthorizeComponent,
+						data: { state: false }
 					},
 					{
-						path: 'reports',
-						component: ReportsComponent,
-						data: { selectors: { project: false } }
+						path: ':id',
+						component: UpworkComponent,
+						children: [
+							{
+								path: '',
+								redirectTo: 'contracts',
+								pathMatch: 'full'
+							},
+							{
+								path: 'activities',
+								component: TransactionsComponent
+							},
+							{
+								path: 'reports',
+								component: ReportsComponent,
+								data: { selectors: { project: false } }
+							},
+							{
+								path: 'transactions',
+								component: TransactionsComponent
+							},
+							{
+								path: 'contracts',
+								component: ContractsComponent
+							}
+						]
 					},
 					{
-						path: 'transactions',
-						component: TransactionsComponent
-					},
-					{
-						path: 'contracts',
-						component: ContractsComponent
+						path: ':id/settings',
+						loadChildren: () => import('@gauzy/ui-core/shared').then((m) => m.WorkInProgressModule)
 					}
 				]
-			},
-			{
-				path: ':id/settings',
-				loadChildren: () => import('@gauzy/ui-core/shared').then((m) => m.WorkInProgressModule)
 			}
 		])
 	],
